spec 021 bf16-α: store recur_alpha buffer in bfloat16, drop runtime cast
Hypothesis: 021's post-loop train-loss regression vs 019b (+0.005 to
+0.023 per step from loop activation onward) is not from α values — the
dc0b5f8 fix of pass-3 L4 0.96484375 → 0.97265625 had zero measurable
effect on training loss. The remaining mechanism candidate is
buffer-vs-literal numerical path:

  - 019b literal-α: Inductor const-folds the Python float into a single
    fused bf16 pointwise kernel.
  - 021 buffer-α (float32): buffer read + .to(bf16) cast + blend — not
    const-foldable, may prevent fusion, introduces extra kernel boundary.

Fix: store buffer as bfloat16 directly and drop the .to() cast at the
two blend sites. All six α values (multiples of 1/128, ≤ 1.5) fit bf16's
7-bit mantissa exactly, so zero precision loss vs float32 storage.

Test plan: rerun seed_42 on this commit. If post-loop losses match 019b
within pod variance, buffer-α viability is restored and 3-seed becomes
live again.

+++ records/track_10min_16mb/2026-04-19_SP8192_CaseOps_GatedAttn_QuantGate_Loop45_PhasedTTT/train_gpt.py
@@ -1054,10 +1054,14 @@ class GPT(nn.Module):
             # suffered (where literal α kernels paid a cold-path penalty after
             # every train→eval→train mode switch at val_loss_every=4000).
             # Expected throughput profile matches 017's tensor-α recipe.
+            # dtype=bfloat16: matches activation dtype so no cast kernel is
+            # needed at blend time. All six α values are exact multiples of
+            # 1/128 (≤ 1.5) — fit bf16's 7-bit mantissa exactly, zero precision
+            # loss vs storing in float32.
             _recur_alpha_017_endpoint = torch.tensor(
                 [[1.078125, 1.2734375, 1.4296875],     # pass-2 L3, L4, L5
                  [1.015625, 0.97265625, 0.83203125]],  # pass-3 L3, L4, L5
-                dtype=torch.float32,
+                dtype=torch.bfloat16,
             )
             assert _recur_alpha_017_endpoint.shape == (h.num_loops, num_looped), (
                 f"017 endpoint table shape {_recur_alpha_017_endpoint.shape} "
@@ -1211,7 +1215,7 @@ class GPT(nn.Module):
             x_new = self.blocks[i](x_before, x0, q_w, k_w, v_w, out_w, up_w, down_w, cu_seqlens=cu_seqlens, max_seqlen=max_seqlen)
             if enc_alpha_info is not None and enc_alpha_info[step_idx] is not None:
                 pass_off, local_idx = enc_alpha_info[step_idx]
-                alpha = self.recur_alpha[pass_off, local_idx].to(x_new.dtype)
+                alpha = self.recur_alpha[pass_off, local_idx]
                 x = alpha * x_new + (1.0 - alpha) * x_before
                 # Diagnostic: p2p cosine similarity on block deltas (optional).
                 if self.recur_diag_p2p_cos:
@@ -1272,7 +1276,7 @@ class GPT(nn.Module):
                 x_new = self.blocks[i](x_before, x0, q_w, k_w, v_w, out_w, up_w, down_w, cu_seqlens=cu_seqlens, max_seqlen=max_seqlen)
                 if dec_alpha_info is not None and dec_alpha_info[skip_idx] is not None:
                     pass_off, local_idx = dec_alpha_info[skip_idx]
-                    alpha = self.recur_alpha[pass_off, local_idx].to(x_new.dtype)
+                    alpha = self.recur_alpha[pass_off, local_idx]
                     x = alpha * x_new + (1.0 - alpha) * x_before
                     if self.recur_diag_p2p_cos:
                         delta_this = (x_new - x_before).detach()
